refactor(components): extract components into separate files
Move AppIconView, ClipboardItemRow, ESCKeyCatcher, HotkeyCapturingView and TextPreview into dedicated files under Views/Components directory to improve code organization and maintainability
